+++ src/windows/tray_icon.rs
@@ -1,9 +1,9 @@
-use muda::Submenu;
+use muda::{PredefinedMenuItem, Submenu};
 use tray_icon::menu::{Menu, MenuItem};
 use tray_icon::TrayIconBuilder;
 use winit::event_loop::{ActiveEventLoop, EventLoopProxy};
 
-use crate::app::AppMessage;
+use crate::windows::app::AppMessage;
 
 pub struct TrayIcon {
     #[allow(unused)]
@@ -11,18 +11,37 @@ pub struct TrayIcon {
     proxy: EventLoopProxy<AppMessage>,
     #[allow(unused)]
     menu: Menu,
-    quit: MenuItem,
     move_resize: Submenu,
     move_resize_items: Vec<MenuItem>,
+    refresh: MenuItem,
+    quit: MenuItem,
 }
 
 impl TrayIcon {
     pub fn new(proxy: EventLoopProxy<AppMessage>) -> anyhow::Result<Self> {
         let icon = tray_icon::Icon::from_resource(1, Some((32, 32)))?;
 
-        let quit = MenuItem::new("Quit", true, None);
         let move_resize = Submenu::new("Move && Resize", true);
-        let menu = Menu::with_items(&[&move_resize, &quit])?;
+        let refresh = MenuItem::new("Refresh", true, None);
+        let separator = PredefinedMenuItem::separator();
+
+        #[cfg(debug_assertions)]
+        let title = MenuItem::new(concat!(env!("CARGO_PKG_NAME"), " (debug)"), false, None);
+        #[cfg(not(debug_assertions))]
+        let title = MenuItem::new(env!("CARGO_PKG_NAME"), false, None);
+
+        let version = MenuItem::new(concat!("v", env!("CARGO_PKG_VERSION")), false, None);
+        let quit = MenuItem::new("Quit", true, None);
+
+        let menu = Menu::with_items(&[
+            &move_resize,
+            &refresh,
+            &separator,
+            &title,
+            &version,
+            &separator,
+            &quit,
+        ])?;
 
         TrayIconBuilder::new()
             .with_icon(icon)
@@ -34,10 +53,15 @@ impl TrayIcon {
                 icon,
                 proxy,
                 menu,
-                quit,
                 move_resize,
                 move_resize_items: vec![],
+                refresh,
+                quit,
             })
+    }
+
+    pub fn is_move_resize_item(&self, id: &str) -> bool {
+        self.move_resize_items.iter().any(|item| item.id() == id)
     }
 
     pub fn destroy_items_for_switchers(&mut self) -> anyhow::Result<()> {
@@ -66,16 +90,14 @@ impl TrayIcon {
         event: &AppMessage,
     ) -> anyhow::Result<()> {
         match event {
-            AppMessage::MenuEvent(event) if event.id() == self.quit.id() => event_loop.exit(),
-            AppMessage::MenuEvent(event)
-                if self
-                    .move_resize_items
-                    .iter()
-                    .any(|item| item.id() == event.id()) =>
-            {
-                self.proxy
-                    .send_event(AppMessage::StartMoveResize(event.id().as_ref().to_string()))?;
+            AppMessage::MenuEvent(event) if self.is_move_resize_item(event.id().0.as_str()) => {
+                let id = event.id().0.clone();
+                self.proxy.send_event(AppMessage::StartMoveResize(id))?;
             }
+            AppMessage::MenuEvent(event) if event.id() == self.refresh.id() => {
+                self.proxy.send_event(AppMessage::RecreateSwitcherWindows)?
+            }
+            AppMessage::MenuEvent(event) if event.id() == self.quit.id() => event_loop.exit(),
             _ => {}
         }